Fix the problem of image normalizaiton first, offset should be set to 127.0f when image_min < 0 second, byte_t can't be transformed to int directly in visualdl/logic/sdk.cc

+++ visualdl/logic/sdk.cc
@@ -272,7 +272,7 @@ ImageReader::ImageRecord ImageReader::record(int offset, int index) {
   std::transform(brcd.data.begin(),
                  brcd.data.end(),
                  std::back_inserter(res.data),
-                 [](byte_t i) { return (int)(i); });
+                 [](byte_t i) { return (uint8_t)(i); });
   res.shape = entry.GetMulti<shape_t>();
   res.step_id = record.id();
   return res;
